pulumi/pyproject.toml
[project]
name = ""
version = "0.0.1"
dependencies = [
    "annotated-types==0.7.0",
    "arpeggio==2.0.2",
    "attrs==24.2.0",
    "certifi==2024.7.4",
    "charset-normalizer==3.3.2",
    "dill==0.3.8",
    "grpcio==1.60.2",
    "idna==3.7",
    "parver==0.5",
    "protobuf==4.25.4",
    "pulumi==3.128.0",
    "pulumi-command==1.0.1",
    "pulumi-kubernetes==4.16.0",
    "pydantic==2.8.2",
    "pydantic-core==2.20.1",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "semver==2.13.0",
    "six==1.16.0",
    "typing-extensions==4.12.2",
    "urllib3==2.2.2",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "types-PyYAML>=6.0.0",
]
